refactor(service): move users and trainers out of service/auth/
service/auth/users.ts and service/auth/trainers.ts handle profile
management, not authentication. Moving them to service/ so the auth/
subdirectory only contains auth concerns (login, signup, verification,
password reset, email, NextAuth config).
